+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Migrate from lake data to fastnear data
 * Add metrics to calculate the number of blocks which fetched from the cache and fastnear
 * Add blocks chunks cache and optimize cache layer for read-rpc-server
+* Add authorization token supports
 
 ## [0.3.3](https://github.com/near/read-rpc/releases/tag/v0.3.3)